+++ src/main/java/com/project/springapistudy/common/exception/DuplicationMenuException.java
@@ -1,4 +1,4 @@
-package com.project.springapistudy.web.exception;
+package com.project.springapistudy.common.exception;
 
 public class DuplicationMenuException extends RuntimeException {
     public DuplicationMenuException(String message) {
