+++ src/main/java/com/project/springapistudy/common/handler/ValidExceptionHandler.java
@@ -1,6 +1,6 @@
-package com.project.springapistudy.web.exception.handler;
+package com.project.springapistudy.common.handler;
 
-import com.project.springapistudy.util.ResponseData;
+import com.project.springapistudy.common.util.ResponseData;
 import org.springframework.context.support.DefaultMessageSourceResolvable;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.BindingResult;
@@ -10,7 +10,6 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @RestControllerAdvice
 public class ValidExceptionHandler {
