+++ src/main/java/com/project/springapistudy/common/util/ResponseData.java
@@ -1,4 +1,4 @@
-package com.project.springapistudy.util;
+package com.project.springapistudy.common.util;
 
 import lombok.Getter;
